requirements.txt
asgiref==3.8.1
asttokens==2.4.1
decorator==5.1.1
Django==4.2.10
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
django-extensions==3.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.2
executing==2.1.0
flake8==7.1.1
ipython==8.27.0
isort==5.13.2
jedi==0.19.1
matplotlib-inline==0.1.7
mccabe==0.7.0
parso==0.8.4
pexpect==4.9.0
prompt_toolkit==3.0.47
ptyprocess==0.7.0
pure_eval==0.2.3
pycodestyle==2.12.1
pyflakes==3.2.0
Pygments==2.18.0
PyJWT==2.9.0
pytz==2024.2
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.12.2
wcwidth==0.2.13
